Add AdjacentNodes.py
Create new projects directory to hold different programming questions.

This problem draws inspiration from
https://edabit.com/challenge/3DAkZHv2LZjgqWbvW.

The goal WAS to check if a node matrix had any adjacent nodes but this
is difficult to check when you have no use cases.

This program generates 2d matrices which can be either directed or
undirected node graphs. There are two functions to create directed
graphs, one which is much faster than the other by taking advantage of
the symmetry in directed adjacent node graphs.